Stop tab completion on no input

+++ terminal.js
@@ -37,6 +37,9 @@ $(document).ready(function(){
   $intro_run=1;
 
   function find_tab_completed_command(command) {
+    if command == ""
+      return "";
+
     var commands = ['help', 'clear', 'reboot'];
     for (var i = 0; i < pages.length; i ++) {
       var page = pages[i];
